Error Handling, Code Cleanup.
Removed print statements. Added error handlers for revcat, firebase
