+++ lib/bloc/cards/card_event.dart
@@ -1,19 +1,17 @@
 part of 'card_bloc.dart';
 
 abstract class CardEvent extends Equatable {
+  Category? category;
   @override
-  // TODO: implement props
-  List<Object?> get props => throw UnimplementedError();
+  List<Object?> get props => [category];
 }
 
 class LoadCards extends CardEvent {
   final Category category;
-
   LoadCards({
     required this.category,
   });
   @override
-  // TODO: implement props
   List<Object?> get props => [category];
 }
