+++ lib/bloc/cards/card_state.dart
@@ -3,12 +3,11 @@ part of 'card_bloc.dart';
 abstract class CardState extends Equatable {
   final Category? category;
   final List<String>? cardImageUrls;
-  CardState({
+  const CardState({
     this.category,
     this.cardImageUrls,
   });
   @override
-  // TODO: implement props
   List<Object?> get props => [category, cardImageUrls];
 }
 
@@ -17,10 +16,12 @@ class CardsInitial extends CardState {}
 class CardsLoading extends CardState {}
 
 class CardsLoaded extends CardState {
+  @override
   final Category category;
+  @override
   final List<String> cardImageUrls;
 
-  CardsLoaded({
+  const CardsLoaded({
     required this.category,
     required this.cardImageUrls,
   });
