+++ lib/bloc/group/group_event.dart
@@ -2,12 +2,20 @@ part of 'group_bloc.dart';
 
 @immutable
 abstract class GroupEvent extends Equatable {
+  String? userId;
+  User? manager;
+  Group? group;
+  GroupEvent({
+    this.userId,
+    this.manager,
+    this.group,
+  });
   @override
-  // TODO: implement props
-  List<Object?> get props => throw UnimplementedError();
+  List<Object?> get props => [userId, manager, group];
 }
 
 class LoadGroups extends GroupEvent {
+  @override
   final String userId;
   LoadGroups({
     required this.userId,
@@ -15,22 +23,25 @@ class LoadGroups extends GroupEvent {
 }
 
 class CreateGroup extends GroupEvent {
-  User manager;
-  Group group;
+  @override
+  final User manager;
+  @override
+  final Group group;
   CreateGroup({
     required this.manager,
     required this.group,
   });
 
   @override
-  // TODO: implement props
   List<Object?> get props => [group, manager];
 }
 
 class EditGroup extends GroupEvent {}
 
 class DeleteGroup extends GroupEvent {
+  @override
   final Group group;
+  @override
   final User manager;
   DeleteGroup({
     required this.group,
@@ -38,20 +49,20 @@ class DeleteGroup extends GroupEvent {
   });
 
   @override
-  // TODO: implement props
   List<Object?> get props => [group, manager];
 }
 
 class UpdateGroup extends GroupEvent {
-  User manager;
-  Group group;
+  @override
+  final User manager;
+  @override
+  final Group group;
   UpdateGroup({
     required this.group,
     required this.manager,
   });
 
   @override
-  // TODO: implement props
   List<Object?> get props => [group];
 }
