+++ lib/data/repositories/storage/storage_repository.dart
@@ -1,5 +1,9 @@
+import 'package:firebase_core/firebase_core.dart';
+import 'package:firebase_crashlytics/firebase_crashlytics.dart';
 import 'package:firebase_storage/firebase_storage.dart' as firebase_storage;
+import 'package:flutter/material.dart';
 import 'package:inspired_senior_care_app/data/repositories/storage/base_storage_repository.dart';
+import 'package:inspired_senior_care_app/globals.dart';
 
 class StorageRepository extends BaseStorageRepository {
   final firebase_storage.FirebaseStorage storage =
@@ -17,19 +21,23 @@ class StorageRepository extends BaseStorageRepository {
       }
 
       return coverUrls;
-    } catch (e) {
-      print(e);
-      throw Exception(e);
+    } on FirebaseException catch (e) {
+      final SnackBar snackBar = SnackBar(
+        content: Text(e.message.toString()),
+        backgroundColor: Colors.redAccent,
+      );
+      snackbarKey.currentState?.showSnackBar(snackBar);
+      (e, stack) =>
+          FirebaseCrashlytics.instance.recordError(e, stack, fatal: true);
+      throw Exception();
     }
   }
 
   Future<List<String>> getCategoryCards(String categoryName) async {
     try {
-      print('Getting Cards...');
       var cardList = await storage.ref('card-contents/$categoryName').listAll();
 
       int cardCount = cardList.items.length;
-      print('This many cards: $cardCount');
       List<String> cardImageURLs = [];
 
       for (int i = 1; i < cardCount + 1; i++) {
@@ -40,9 +48,15 @@ class StorageRepository extends BaseStorageRepository {
       }
 
       return cardImageURLs;
-    } catch (e) {
-      print('Error Loading Cards!');
-      throw Exception(e);
+    } on FirebaseException catch (e) {
+      final SnackBar snackBar = SnackBar(
+        content: Text(e.message.toString()),
+        backgroundColor: Colors.redAccent,
+      );
+      snackbarKey.currentState?.showSnackBar(snackBar);
+      (e, stack) =>
+          FirebaseCrashlytics.instance.recordError(e, stack, fatal: true);
+      throw Exception();
     }
   }
 }
