+++ lib/view/pages/main/manager_share_deck_page.dart
@@ -3,7 +3,6 @@ import 'package:flutter/material.dart';
 import 'package:flutter_animate/flutter_animate.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:font_awesome_flutter/font_awesome_flutter.dart';
-import 'package:go_router/go_router.dart';
 import 'package:infinite_carousel/infinite_carousel.dart';
 import 'package:inspired_senior_care_app/bloc/cards/card_bloc.dart';
 import 'package:inspired_senior_care_app/bloc/deck/deck_cubit.dart';
@@ -12,10 +11,10 @@ import 'package:inspired_senior_care_app/bloc/purchases/purchases_bloc.dart';
 import 'package:inspired_senior_care_app/bloc/share_bloc/share_bloc.dart';
 import 'package:inspired_senior_care_app/data/models/category.dart';
 import 'package:inspired_senior_care_app/data/models/user.dart';
+import 'package:inspired_senior_care_app/view/pages/IAP/premium_org_dialog.dart';
 import 'package:inspired_senior_care_app/view/widget/main/bottom_app_bar.dart';
 import 'package:loading_animation_widget/loading_animation_widget.dart';
 import 'package:purchases_flutter/models/offerings_wrapper.dart';
-import 'package:purchases_flutter/models/package_wrapper.dart';
 
 class ManagerShareDeckPage extends StatefulWidget {
   final Category category;
@@ -62,9 +61,7 @@ class _ManagerShareDeckPageState extends State<ManagerShareDeckPage> {
         preferredSize: const Size.fromHeight(50),
         child: BlocConsumer<DeckCubit, DeckState>(
           listener: (context, state) {
-            if (state.status == DeckStatus.zoomed) {
-              print('Deck Zoomed');
-            }
+            if (state.status == DeckStatus.zoomed) {}
           },
           builder: (context, state) {
             if (state.status == DeckStatus.zoomed) {
@@ -113,8 +110,6 @@ class _ManagerShareDeckPageState extends State<ManagerShareDeckPage> {
               context.watch<PurchasesBloc>().state.isSubscribed;
           //  int currentCard = context.watch<DeckCubit>().currentCardNumber;
 
-          print('Card Number: $currentCard');
-
           if (state is PurchasesLoading) {
             return Center(
               child: LoadingAnimationWidget.discreteCircle(
@@ -161,7 +156,6 @@ class _ManagerShareDeckPageState extends State<ManagerShareDeckPage> {
                     if (currentCard - 1 >=
                             (widget.category.totalCards! / 2).round() &&
                         (isSubscribed == false || isSubscribed == null)) {
-                      print('Is Subscribed: $isSubscribed');
                       WidgetsBinding.instance
                           .addPostFrameCallback((_) => showDialog(
                                 barrierDismissible: false,
@@ -528,580 +522,6 @@ class DeckCompleteDialog extends StatelessWidget {
           ],
         ),
       ),
-    );
-  }
-}
-
-class PremiumIndividualOfferDialog extends StatefulWidget {
-  const PremiumIndividualOfferDialog({
-    Key? key,
-  }) : super(key: key);
-
-  @override
-  State<PremiumIndividualOfferDialog> createState() =>
-      _PremiumOfferDialogState();
-}
-
-double borderTopOffset = 65;
-
-class _PremiumOfferDialogState extends State<PremiumIndividualOfferDialog> {
-  String? selectedOffer;
-
-  @override
-  Widget build(BuildContext context) {
-    return BlocConsumer<PurchasesBloc, PurchasesState>(
-      listener: (context, state) {
-        if (state is PurchasesUpdated) {
-          Navigator.pop(context);
-        }
-      },
-      buildWhen: (previous, current) =>
-          previous.selectedPackage != current.selectedPackage,
-      builder: (context, state) {
-        if (state is PurchasesLoading) {
-          return LoadingAnimationWidget.inkDrop(color: Colors.blue, size: 30.0);
-        }
-        if (state is PurchasesLoaded) {
-          List<Package?> packages = [];
-
-          return Dialog(
-            backgroundColor: Colors.grey.shade100,
-            shape:
-                RoundedRectangleBorder(borderRadius: BorderRadius.circular(20)),
-            child: Padding(
-              padding:
-                  const EdgeInsets.symmetric(vertical: 36.0, horizontal: 16.0),
-              child: BlocBuilder<PurchasesBloc, PurchasesState>(
-                builder: (context, state) {
-                  if (state is PurchasesLoading) {
-                    return LoadingAnimationWidget.inkDrop(
-                        color: Colors.blue, size: 30.0);
-                  }
-                  if (state is PurchasesLoaded) {
-                    if (context.watch<ProfileBloc>().state.user.type ==
-                        'user') {
-                      packages.addAll([
-                        state.offerings
-                            ?.getOffering('subscriptions')!
-                            .getPackage('Individual (Monthly)'),
-                        state.offerings
-                            ?.getOffering('subscriptions')!
-                            .getPackage('Individual (Yearly)'),
-                      ]);
-                      print('Got ${packages.length} Offerings');
-                    } else if (context.watch<ProfileBloc>().state.user.type ==
-                        'manager') {
-                      packages.addAll([
-                        state.offerings
-                            ?.getOffering('subscriptions')!
-                            .getPackage('Organization (Monthly)'),
-                        state.offerings
-                            ?.getOffering('subscriptions')!
-                            .getPackage('Organization (Annual)')
-                      ]);
-                    }
-
-                    return Stack(
-                      clipBehavior: Clip.none,
-                      alignment: AlignmentDirectional.topEnd,
-                      children: [
-                        Column(mainAxisSize: MainAxisSize.min, children: [
-                          Padding(
-                            padding:
-                                const EdgeInsets.only(top: 12.0, bottom: 8.0),
-                            child: Text(
-                              'Upgrade To Keep Going!',
-                              style: Theme.of(context).textTheme.headlineSmall,
-                            ),
-                          ),
-                          SizedBox(
-                              width: 325,
-                              height: 150,
-                              child: Padding(
-                                padding:
-                                    const EdgeInsets.symmetric(vertical: 4.0),
-                                child: IndividualOfferCard(
-                                  packages: packages,
-                                ),
-                              )),
-                          Padding(
-                            padding: const EdgeInsets.only(
-                                left: 24.0, right: 24.0, bottom: 8.0),
-                            child: ElevatedButton(
-                                onPressed: () {
-                                  Package? package = context
-                                      .read<PurchasesBloc>()
-                                      .selectedPackage;
-
-                                  context.read<PurchasesBloc>().add(AddPurchase(
-                                      package: package ?? packages[0]!));
-
-                                  // Navigator.pop(context);
-                                },
-                                child: const Text('Subscribe')),
-                          ),
-                          Padding(
-                            padding:
-                                const EdgeInsets.symmetric(horizontal: 24.0),
-                            child: Text(
-                              'Subscribe & instantly gain access to the rest of the cards!',
-                              style: Theme.of(context).textTheme.titleMedium,
-                              textAlign: TextAlign.center,
-                            ),
-                          ),
-                        ]),
-                        Positioned(
-                          top: -25,
-                          right: -5,
-                          child: SizedBox(
-                            height: 40,
-                            child: CloseButton(
-                              onPressed: () {
-                                Navigator.popUntil(
-                                    context, ModalRoute.withName('categories'));
-                              },
-                            ),
-                          ),
-                        ),
-                      ],
-                    );
-                  } else {
-                    return const Center(
-                      child: Text('Something Went Wrong...'),
-                    );
-                  }
-                },
-              ),
-            ),
-          );
-        } else {
-          return const Center(
-            child: Text('Something Went Wrong...'),
-          );
-        }
-      },
-    );
-  }
-}
-
-class PremiumOrganizationOfferDialog extends StatefulWidget {
-  const PremiumOrganizationOfferDialog({
-    Key? key,
-  }) : super(key: key);
-  @override
-  State<PremiumOrganizationOfferDialog> createState() =>
-      _PremiumOrganizationOfferDialogState();
-}
-
-class _PremiumOrganizationOfferDialogState
-    extends State<PremiumOrganizationOfferDialog> {
-  String? selectedOffer;
-  @override
-  Widget build(BuildContext context) {
-    return BlocConsumer<PurchasesBloc, PurchasesState>(
-      listener: (context, state) {
-        if (state is ProfileUpdated) {
-          Navigator.pop(context);
-        }
-      },
-      builder: (context, state) {
-        if (state is PurchasesLoading) {
-          return LoadingAnimationWidget.inkDrop(color: Colors.blue, size: 30.0);
-        }
-        if (state is PurchasesLoaded) {
-          List<Package?> packages = [];
-          return Dialog(
-            backgroundColor: Colors.grey.shade100,
-            shape:
-                RoundedRectangleBorder(borderRadius: BorderRadius.circular(20)),
-            child: Padding(
-              padding:
-                  const EdgeInsets.symmetric(vertical: 36.0, horizontal: 16.0),
-              child: BlocBuilder<PurchasesBloc, PurchasesState>(
-                builder: (context, state) {
-                  if (state is PurchasesLoading) {
-                    return LoadingAnimationWidget.inkDrop(
-                        color: Colors.blue, size: 30.0);
-                  }
-                  if (state is PurchasesLoaded) {
-                    if (context.watch<ProfileBloc>().state.user.type ==
-                        'user') {
-                      packages.addAll([
-                        state.offerings
-                            ?.getOffering('subscriptions')!
-                            .getPackage('Individual (Monthly)'),
-                        state.offerings
-                            ?.getOffering('subscriptions')!
-                            .getPackage('Individual (Yearly)'),
-                      ]);
-                      print('Got ${packages.length} Offerings');
-                    } else if (context.watch<ProfileBloc>().state.user.type ==
-                        'manager') {
-                      packages.addAll([
-                        state.offerings
-                            ?.getOffering('subscriptions')!
-                            .getPackage('Organization (Monthly)'),
-                        state.offerings
-                            ?.getOffering('subscriptions')!
-                            .getPackage('Organization (Annual)')
-                      ]);
-                    }
-                    return Stack(
-                      clipBehavior: Clip.none,
-                      alignment: AlignmentDirectional.topEnd,
-                      children: [
-                        Column(mainAxisSize: MainAxisSize.min, children: [
-                          Padding(
-                            padding:
-                                const EdgeInsets.only(top: 12.0, bottom: 8.0),
-                            child: Text(
-                              'Upgrade Organization!',
-                              style: Theme.of(context).textTheme.headlineSmall,
-                            ),
-                          ),
-                          SizedBox(
-                              width: 325,
-                              height: 160,
-                              child: Padding(
-                                padding:
-                                    const EdgeInsets.symmetric(vertical: 4.0),
-                                child: OrganizationOfferCard(
-                                  packages: packages,
-                                ),
-                              )),
-                          Padding(
-                            padding: const EdgeInsets.only(
-                                left: 24.0, right: 24.0, bottom: 8.0),
-                            child: ElevatedButton(
-                                onPressed: () {
-                                  Package? package = context
-                                      .read<PurchasesBloc>()
-                                      .selectedPackage;
-                                  context.read<PurchasesBloc>().add(AddPurchase(
-                                      package: package ?? packages[0]!));
-                                  // Navigator.pop(context);
-                                },
-                                child: const Text('Subscribe')),
-                          ),
-                          Padding(
-                            padding:
-                                const EdgeInsets.symmetric(horizontal: 24.0),
-                            child: Text(
-                              'Subscribe & instantly gain access to the rest of the cards!',
-                              style: Theme.of(context).textTheme.titleMedium,
-                              textAlign: TextAlign.center,
-                            ),
-                          ),
-                        ]),
-                        Positioned(
-                          top: -25,
-                          right: -5,
-                          child: SizedBox(
-                            height: 40,
-                            child: CloseButton(
-                              onPressed: () {
-                                context.pop();
-                              },
-                            ),
-                          ),
-                        ),
-                      ],
-                    );
-                  } else {
-                    return const Center(
-                      child: Text('Something Went Wrong...'),
-                    );
-                  }
-                },
-              ),
-            ),
-          );
-        } else {
-          return const Center(
-            child: Text('Something Went Wrong...'),
-          );
-        }
-      },
-    );
-  }
-}
-
-int selectedOffer = 0;
-String? selectedTerm;
-
-class IndividualOfferCard extends StatefulWidget {
-  final List<Package?> packages;
-  const IndividualOfferCard({super.key, required this.packages});
-
-  @override
-  State<IndividualOfferCard> createState() => _IndividualOfferCardState();
-}
-
-List<String> menuItemList = ['Monthly', 'Yearly'];
-
-class _IndividualOfferCardState extends State<IndividualOfferCard> {
-  String dropdownValue = menuItemList.first;
-  @override
-  Widget build(BuildContext context) {
-    return BlocBuilder<PurchasesBloc, PurchasesState>(
-      builder: (context, state) {
-        if (state is PurchasesLoading) {
-          return LoadingAnimationWidget.inkDrop(color: Colors.blue, size: 30.0);
-        }
-        if (state is PurchasesLoaded) {
-          return Column(mainAxisSize: MainAxisSize.min, children: [
-            SizedBox(
-              height: 140,
-              width: 280,
-              child: Card(
-                shape: RoundedRectangleBorder(
-                  borderRadius: BorderRadius.circular(16.0),
-                ),
-                elevation: 1.618,
-                child: Column(
-                  mainAxisSize: MainAxisSize.min,
-                  mainAxisAlignment: MainAxisAlignment.center,
-                  children: [
-                    Padding(
-                      padding: const EdgeInsets.symmetric(
-                          horizontal: 18.0, vertical: 8.0),
-                      child: Row(
-                        mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                        children: [
-                          Text(
-                            'Individual',
-                            style: Theme.of(context).textTheme.titleLarge,
-                          ),
-                          Row(
-                            mainAxisAlignment: MainAxisAlignment.end,
-                            children: [
-                              if (dropdownValue == 'Monthly')
-                                Text(
-                                  '${widget.packages[0]?.storeProduct.priceString}/mo.',
-                                  style:
-                                      Theme.of(context).textTheme.titleMedium,
-                                ),
-                              if (dropdownValue == 'Yearly')
-                                Text(
-                                  '${widget.packages[1]?.storeProduct.priceString}/yr.',
-                                  style:
-                                      Theme.of(context).textTheme.titleMedium,
-                                ),
-                            ],
-                          ),
-                        ],
-                      ),
-                    ),
-                    ListTile(
-                      leading: CircleAvatar(
-                          radius: 18,
-                          backgroundColor: Colors.orange.shade800,
-                          child: const Icon(
-                            Icons.person_rounded,
-                            color: Colors.white,
-                          )),
-                      subtitle: const Text(
-                          'Unlock access to all cards & categories.'),
-                    ),
-                    Padding(
-                      padding: const EdgeInsets.only(right: 16.0, bottom: 8.0),
-                      child: Row(
-                        mainAxisAlignment: MainAxisAlignment.end,
-                        children: [
-                          DropdownButton<String>(
-                            isDense: true,
-                            underline: const SizedBox(),
-                            borderRadius: BorderRadius.circular(20),
-                            value: dropdownValue,
-                            items: menuItemList
-                                .map<DropdownMenuItem<String>>((String value) {
-                              return DropdownMenuItem<String>(
-                                  value: value,
-                                  child: Chip(
-                                    padding: const EdgeInsets.all(4.0),
-                                    backgroundColor: Colors.blueAccent,
-                                    visualDensity: VisualDensity.compact,
-                                    label: Padding(
-                                        padding:
-                                            const EdgeInsets.only(bottom: 4.0),
-                                        child: Text(
-                                          value,
-                                          style: const TextStyle(
-                                            color: Colors.white,
-                                            fontSize: 12.0,
-                                          ),
-                                        )),
-                                  ));
-                            }).toList(),
-                            onChanged: (String? value) {
-                              // This is called when the user selects an item.
-                              setState(() {
-                                dropdownValue = value!;
-                              });
-                              if (dropdownValue == 'Monthly') {
-                                context.read<PurchasesBloc>().add(SelectPackage(
-                                    package: widget.packages[0]!));
-                                print('Monthly Selected');
-                              } else if (dropdownValue == 'Yearly') {
-                                context.read<PurchasesBloc>().add(SelectPackage(
-                                    package: widget.packages[1]!));
-                                print('Yearly Selected');
-                              }
-                            },
-                          ),
-                        ],
-                      ),
-                    ),
-                  ],
-                ),
-              ),
-            )
-          ]);
-        } else {
-          return const Center(
-            child: Text('Something Went Wrong..'),
-          );
-        }
-      },
-    );
-  }
-}
-
-class OrganizationOfferCard extends StatefulWidget {
-  final List<Package?> packages;
-  const OrganizationOfferCard({super.key, required this.packages});
-
-  @override
-  State<OrganizationOfferCard> createState() => _OrganizationOfferCardState();
-}
-
-List<String> menuItemList2 = ['Monthly', 'Yearly'];
-
-class _OrganizationOfferCardState extends State<OrganizationOfferCard> {
-  String dropdownValue = menuItemList.first;
-  @override
-  Widget build(BuildContext context) {
-    return BlocBuilder<PurchasesBloc, PurchasesState>(
-      builder: (context, state) {
-        if (state is PurchasesLoading) {
-          return LoadingAnimationWidget.inkDrop(color: Colors.blue, size: 30.0);
-        }
-        if (state is PurchasesLoaded) {
-          return Column(mainAxisSize: MainAxisSize.min, children: [
-            SizedBox(
-              height: 145,
-              width: 280,
-              child: Card(
-                shape: RoundedRectangleBorder(
-                  borderRadius: BorderRadius.circular(16.0),
-                ),
-                elevation: 1.618,
-                child: Column(
-                  mainAxisSize: MainAxisSize.min,
-                  mainAxisAlignment: MainAxisAlignment.center,
-                  children: [
-                    Padding(
-                      padding: const EdgeInsets.symmetric(
-                          horizontal: 18.0, vertical: 8.0),
-                      child: Row(
-                        mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                        children: [
-                          Text(
-                            'Organization',
-                            style: Theme.of(context).textTheme.titleLarge,
-                          ),
-                          Row(
-                            mainAxisAlignment: MainAxisAlignment.end,
-                            children: [
-                              if (dropdownValue == 'Monthly')
-                                Text(
-                                  '${widget.packages[0]?.storeProduct.priceString}/mo.',
-                                  style:
-                                      Theme.of(context).textTheme.titleMedium,
-                                ),
-                              if (dropdownValue == 'Yearly')
-                                Text(
-                                  '${widget.packages[1]?.storeProduct.priceString}/yr.',
-                                  style:
-                                      Theme.of(context).textTheme.titleMedium,
-                                ),
-                            ],
-                          ),
-                        ],
-                      ),
-                    ),
-                    ListTile(
-                      leading: CircleAvatar(
-                          radius: 18,
-                          backgroundColor: Colors.orange.shade800,
-                          child: const Icon(
-                            Icons.person_rounded,
-                            color: Colors.white,
-                          )),
-                      subtitle: const Text(
-                          'Unlock full access for you & all group members.'),
-                    ),
-                    Padding(
-                      padding: const EdgeInsets.only(
-                          right: 16.0, bottom: 8.0, top: 8.0),
-                      child: Row(
-                        mainAxisAlignment: MainAxisAlignment.end,
-                        children: [
-                          DropdownButton<String>(
-                            isDense: true,
-                            underline: const SizedBox(),
-                            borderRadius: BorderRadius.circular(20),
-                            value: dropdownValue,
-                            items: menuItemList2
-                                .map<DropdownMenuItem<String>>((String value) {
-                              return DropdownMenuItem<String>(
-                                  value: value,
-                                  child: Chip(
-                                    padding: const EdgeInsets.all(4.0),
-                                    backgroundColor: Colors.blueAccent,
-                                    visualDensity: VisualDensity.compact,
-                                    label: Padding(
-                                        padding:
-                                            const EdgeInsets.only(bottom: 4.0),
-                                        child: Text(
-                                          value,
-                                          style: const TextStyle(
-                                            color: Colors.white,
-                                            fontSize: 12.0,
-                                          ),
-                                        )),
-                                  ));
-                            }).toList(),
-                            onChanged: (String? value) {
-                              // This is called when the user selects an item.
-                              setState(() {
-                                dropdownValue = value!;
-                              });
-                              if (dropdownValue == 'Monthly') {
-                                context.read<PurchasesBloc>().add(SelectPackage(
-                                    package: widget.packages[0]!));
-                                print('Monthly Selected');
-                              } else if (dropdownValue == 'Yearly') {
-                                context.read<PurchasesBloc>().add(SelectPackage(
-                                    package: widget.packages[1]!));
-                                print('Yearly Selected');
-                              }
-                            },
-                          ),
-                        ],
-                      ),
-                    ),
-                  ],
-                ),
-              ),
-            )
-          ]);
-        } else {
-          return const Center(
-            child: Text('Something Went Wrong..'),
-          );
-        }
-      },
     );
   }
 }
